Altered RepReader to use local html file
Due to the way the reputation system on care-tags.org works, you must be logged in to view a user's reputation. This means that the script would need to be logged in somehow (possible todo?).

For now, I've changed it so that it reads everything from the local file "rep.html". Still need to finish parsing it correctly.