Correctly parses the html and creates a list
Repsection consisted of correct values interspaced with "\n" values, which threw off the .children method. These "\n"s are now removed. User and rep are parsed for each rep and a dictionary matches each user to their reputation handed out. The dictionary is then converted into a list with the values aggregated.